Fixes #132
The original code seems to check whether the type is really a
type constructor thus starting with an upper case letter.
The problem is that this check is done on the transformed name
of the type, which may no longer be an upper case letter
(e.g. when one uses camelTo2 '_'). Therefore, the check should
be done on the original name of the type, which then fixes
issue #132.