Support for additionalProperties with Bool value
According to the spec ([here](https://tools.ietf.org/html/draft-wright-json-schema-validation-00#section-5.18)), `additionalProperties` can be `true` (meaning additional properties with any values) or `false` (meaning the same thing as the attribute not being present, I believe) or a schema.

This commit adds a new type, `AdditionalProperties`, with custom JSON encoding to deal with that, and what appears to be reasonable validation for it.